backend/requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
sqlalchemy>=2.0,<3.0
psycopg[binary]>=3.1,<4.0
pydantic>=2.7,<3.0
pydantic-settings>=2.2,<3.0
python-dotenv>=1.0,<2.0
httpx>=0.27,<1.0
python-multipart>=0.0.9,<1.0

# Web scraping / browser automation
# If you install any `scrapling[...]` extras, make sure you run `scrapling install`
# to install browser dependencies.
scrapling[fetchers]>=0.4.7,<0.5